refactor: extract the example app codegen patch function
