+++ packages/create-react-native-library/src/index.ts
@@ -13,6 +13,7 @@ import generateExampleApp, {
 } from './utils/generateExampleApp';
 import { spawn } from './utils/spawn';
 import { version } from '../package.json';
+import { patchExampleAppCodegen } from './utils/patchExampleAppCodegen';
 
 const FALLBACK_BOB_VERSION = '0.25.0';
 
@@ -284,37 +285,6 @@ const args: Record<ArgName, yargs.Options> = {
     choices: EXAMPLE_CHOICES.map(({ value }) => value),
   },
 };
-
-const GRADLE_INVOKE_CODEGEN_TASK = `
-def isNewArchitectureEnabled() {
-  return rootProject.hasProperty("newArchEnabled") && rootProject.getProperty("newArchEnabled") == "true"
-}
-
-if (isNewArchitectureEnabled()) {
-    // Since our library doesn't invoke codegen automatically we need to do it here.
-    tasks.register('invokeLibraryCodegen', Exec) {
-        workingDir "$rootDir/../../"
-        commandLine "yarn", "codegen"
-    }
-    preBuild.dependsOn invokeLibraryCodegen
-}`;
-
-const XCODE_INVOKE_CODEGEN_ACTION = `
-      <PreActions>
-         <ExecutionAction
-            ActionType = "Xcode.IDEStandardExecutionActionsCore.ExecutionActionType.ShellScriptAction">
-            <ActionContent
-               title = "Invoke Codegen"
-               scriptText = "cd &quot;$WORKSPACE_PATH/../../../&quot; &amp;&amp; yarn codegen&#10;">
-            </ActionContent>
-         </ExecutionAction>
-      </PreActions>`;
-
-const PODSPEC_INVOKE_CODEGEN_SCRIPT = `
-  pre_install do |installer|
-    system("cd ../../ && yarn codegen")
-  end
-`;
 
 // FIXME: fix the type
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -847,58 +817,7 @@ async function create(_argv: yargs.Arguments<any>) {
     }
 
     if (arch !== 'legacy') {
-      const appBuildGradlePath = path.join(
-        folder,
-        'example',
-        'android',
-        'app',
-        'build.gradle'
-      );
-      let appBuildGradle = (await fs.readFile(appBuildGradlePath)).toString();
-      appBuildGradle += GRADLE_INVOKE_CODEGEN_TASK;
-
-      await fs.writeFile(appBuildGradlePath, appBuildGradle);
-
-      const exampleAppBuildSchemePath = path.join(
-        folder,
-        'example',
-        'ios',
-        `${options.project.name}Example.xcodeproj`,
-        'xcshareddata',
-        'xcschemes',
-        `${options.project.name}Example.xcscheme`
-      );
-
-      // Add a prebuild action to invoke codegen
-      const exampleAppBuildScheme = (
-        await fs.readFile(exampleAppBuildSchemePath)
-      )
-        .toString()
-        .split('\n');
-      // Used XCode to determine where it inserts the actions
-      const actionTargetLineIndex = exampleAppBuildScheme.findIndex((line) =>
-        line.includes('<BuildActionEntries>')
-      );
-      exampleAppBuildScheme.splice(
-        actionTargetLineIndex,
-        0,
-        XCODE_INVOKE_CODEGEN_ACTION
-      );
-
-      await fs.writeFile(
-        exampleAppBuildSchemePath,
-        exampleAppBuildScheme.join('\n')
-      );
-
-      // Add a preinstall action to the podfile that invokes codegen
-      const podfilePath = path.join(folder, 'example', 'ios', 'Podfile');
-      const podfile = (await fs.readFile(podfilePath)).toString().split('\n');
-      const podfilePostInstallIndex = podfile.findIndex((line) =>
-        line.includes('post_install do |installer|')
-      );
-      podfile.splice(podfilePostInstallIndex, 0, PODSPEC_INVOKE_CODEGEN_SCRIPT);
-
-      await fs.writeFile(podfilePath, podfile.join('\n'));
+      patchExampleAppCodegen(folder, options.project.name);
     }
   }
